+++ src/models/stores/supports.ts
@@ -3,7 +3,7 @@ import { cloneDeep } from "lodash";
 import { InvestigationModelType } from "../curriculum/investigation";
 import { ProblemModelType } from "../curriculum/problem";
 import { getSectionInitials, getSectionTitle, kAllSectionType, SectionType } from "../curriculum/section";
-import { ESupportType, SupportModel, SupportModelType } from "../curriculum/support";
+import { ESupportMode, ESupportType, SupportModel, SupportModelType } from "../curriculum/support";
 import { UnitModelType } from "../curriculum/unit";
 import { DB } from "../../lib/db";
 import { IDocumentProperties } from "../../lib/db-types";
@@ -80,6 +80,13 @@ export const TeacherSupportModel = types
     deleted: false
   })
   .views(self => ({
+    // excludes sticky notes
+    get isTeacherSupport() {
+      return self.support.mode !== ESupportMode.stickyNote;
+    },
+    get isStickyNote() {
+      return self.support.mode === ESupportMode.stickyNote;
+    },
     get sectionTarget() {
       return (self.type === SupportTarget.section) && self.sectionId
               ? self.sectionId
@@ -104,6 +111,9 @@ export const TeacherSupportModel = types
   .actions(self => ({
     setVisible(visible: boolean) {
       self.visible = visible;
+    },
+    setDeleted(deleted: boolean) {
+      self.deleted = deleted;
     }
   }));
 
@@ -135,6 +145,18 @@ export const CurricularSupportModel = types
 
 export const SupportItemModel = types.union(TeacherSupportModel, CurricularSupportModel);
 
+// utility function which finds authored teacher supports and sticky notes
+function hasNewTeacherSupports(teacherSupports: TeacherSupportModelType[], afterTimestamp?: number) {
+  if (afterTimestamp) {
+    const latestAuthoredTime = teacherSupports.reduce((latest, support) => {
+      return (support.authoredTime > latest) ? support.authoredTime : latest;
+    }, 0);
+    return latestAuthoredTime > afterTimestamp;
+  } else {
+    return teacherSupports.length > 0;
+  }
+}
+
 export const SupportsModel = types
   .model("Supports", {
     curricularSupports: types.array(CurricularSupportModel),
@@ -143,10 +165,33 @@ export const SupportsModel = types
     userSupports: types.array(TeacherSupportModel)
   })
   .views((self) => ({
-    get teacherSupports() {
+    // includes standard teacher supports and sticky notes
+    get allTeacherSupports() {
       return self.classSupports
         .concat(self.groupSupports)
-        .concat(self.userSupports);
+        .concat(self.userSupports)
+        .filter((support) => !support.deleted);
+    }
+  }))
+  .views((self) => ({
+    // standard supports (as opposed to sticky notes)
+    get teacherSupports() {
+      return self.allTeacherSupports.filter((support) => support.isTeacherSupport);
+    },
+    get teacherStickyNotes() {
+      return self.allTeacherSupports.filter((support) => support.isStickyNote);
+    }
+  }))
+  .views((self) => ({
+    getTeacherSupportsForUserProblem(target: ISupportTarget): SupportItemModelType[] {
+      return self.teacherSupports.filter(support => {
+        return support.showForUserProblem(target);
+      });
+    },
+    getStickyNotesForUserProblem(target: ISupportTarget): SupportItemModelType[] {
+      return self.teacherStickyNotes.filter(support => {
+        return support.showForUserProblem(target);
+      });
     }
   }))
   .views((self) => ({
@@ -162,26 +207,15 @@ export const SupportsModel = types
         const supports: SupportItemModelType[] = self.curricularSupports.filter((support) => {
           return sectionId ? support.sectionId === sectionId : true;
         });
-
-        const teacherSupports = self.teacherSupports.filter(support => {
-          return support.showForUserProblem(target);
-        });
-
-        return supports.concat(teacherSupports);
+        return supports.concat(self.getTeacherSupportsForUserProblem(target));
     }
   }))
   .views((self) => ({
-    hasNewSupports(afterTimestamp?: number) {
-      if (afterTimestamp) {
-        // true if there has been a teacher support after the last support tab open
-        const latestAuthoredTime = self.teacherSupports.reduce((latest, support) => {
-          return support.authoredTime > latest ? support.authoredTime : latest;
-        }, 0);
-        return latestAuthoredTime > afterTimestamp;
-      } else {
-        // have not opened supports yet so true if any exist
-        return self.allSupports.length > 0;
-      }
+    hasNewTeacherSupports(afterTimestamp?: number) {
+      return hasNewTeacherSupports(self.teacherSupports, afterTimestamp);
+    },
+    hasNewStickyNotes(afterTimestamp?: number) {
+      return hasNewTeacherSupports(self.teacherStickyNotes, afterTimestamp);
     }
   }))
   .actions((self) => {
@@ -193,7 +227,7 @@ export const SupportsModel = types
           return (support: SupportModelType) => {
             supports.push(CurricularSupportModel.create({
               support: cloneDeep(support),
-              type: type as SupportTarget,
+              type,
               sectionId
             }));
           };
@@ -276,9 +310,17 @@ function getSupportCaption(support: UnionSupportModelType, index: number,
 
 export function addSupportDocumentsToStore(params: ICreateFromUnitParams) {
   const { db, documents, investigation, problem, supports, onDocumentCreated } = params;
+  if (!documents) {
+    return;
+  }
   let index = 0;
   let lastSection: string | undefined;
   supports && supports.forEach(async (support: UnionSupportModelType) => {
+    // skip sticky notes
+    if ((support.supportType === SupportType.teacher) && support.isStickyNote) {
+      return;
+    }
+
     const { sectionId } = support;
     if (sectionId === lastSection) {
       ++index;
@@ -308,19 +350,29 @@ export function addSupportDocumentsToStore(params: ICreateFromUnitParams) {
       }
     }
 
-    const content = await getDocumentContentForSupport(support.support, db);
-    if (content) {
-      const document = DocumentModel.create({
-                        uid: "curriculum",
-                        type: SupportPublication,
-                        key: supportKey,
-                        originDoc,
-                        properties,
-                        createdAt: Date.now(),
-                        content: getSnapshot(content)
-                      });
-      documents && documents.add(document);
-      onDocumentCreated && onDocumentCreated(support, document);
+    let document = documents.getDocument(supportKey);
+    if (document) {
+      // update existing document properties if a document exists
+      if (support.supportType === SupportType.teacher) {
+        support.setDeleted(!!properties.isDeleted);
+      }
+      document.setProperties(properties);
+    }
+    else {
+      const content = await getDocumentContentForSupport(support.support, db);
+      if (content) {
+        document = DocumentModel.create({
+                     uid: "curriculum",
+                     type: SupportPublication,
+                     key: supportKey,
+                     originDoc,
+                     properties,
+                     createdAt: Date.now(),
+                     content: getSnapshot(content)
+                   });
+        documents.add(document);
+        onDocumentCreated?.(support, document);
+      }
     }
   });
 }