+++ src/models/stores/workspace.ts
@@ -3,6 +3,7 @@ import { DocumentModelType, DocumentModel } from "../document/document";
 import { SectionModelType } from "../curriculum/section";
 import { LogEventName, Logger } from "../../lib/logger";
 import { kGroupVirtualDocumentType, GroupVirtualDocument } from "../document/group-virtual-document";
+
 export const ProblemWorkspace = "problem";
 export const LearningLogWorkspace = "learningLog";