+++ CHANGELOG.md
@@ -1,5 +1,43 @@
 # Changelog
 
+## Version 1.0.2 - released November 22, 2019
+
+- Bug fixes
+
+### Asset Sizes
+
+| File | Size | % Change from Previous Release |
+|---|---|---|
+| index.css | 499,791 bytes | 0.0% |
+| index.js | 4,365,952 bytes | 0.0% |
+
+## Version 1.0.1 - released November 20, 2019
+
+- Improved teacher dashboard logging
+- Curriculum content updates
+
+### Asset Sizes
+
+| File | Size | % Change from Previous Release |
+|---|---|---|
+| index.css | 499,791 bytes | 0.0% |
+| index.js | 4,365,755 bytes | 0.0% |
+
+## Version 1.0.0 - released November 19, 2019
+
+- Document supports selection of multiple tiles
+- Drag/drop multiple tiles in document
+- Teacher can publish document as support
+- Teacher dashboard styling improvements
+- Teacher can send "sticky notes" to individual students/groups
+
+### Asset Sizes
+
+| File | Size | % Change from Previous Release |
+|---|---|---|
+| index.css | 499,791 bytes | 1.8% |
+| index.js | 4,365,192 bytes | 2.2% |
+
 ## Version 0.10.0 - released November 8, 2019
 
 - Support linked selection for linked geometry/table points